refactor: 배치 내부로 스케줄링 로직 경로 수정
