+++ src/main/java/com/fourthread/ozang/module/domain/weather/batch/scheduler/WeatherCacheWarmupScheduler.java
@@ -1,4 +1,4 @@
-package com.fourthread.ozang.module.domain.weather.scheduler;
+package com.fourthread.ozang.module.domain.weather.batch.scheduler;
 
 import java.time.ZoneId;
 import lombok.RequiredArgsConstructor;
