+++ src/main/java/com/fourthread/ozang/module/domain/weather/batch/scheduler/WeatherDataCleanupScheduler.java
@@ -1,4 +1,4 @@
-package com.fourthread.ozang.module.domain.weather.scheduler;
+package com.fourthread.ozang.module.domain.weather.batch.scheduler;
 
 import java.time.ZoneId;
 import lombok.RequiredArgsConstructor;
@@ -14,53 +14,53 @@ import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
 /**
- * 날씨 변화 감지 스케줄러
- * 활성 지역의 날씨 변화를 감지하고 알림 전송
+ * 날씨 데이터 정리 스케줄러
+ * 오래된 날씨 데이터를 주기적으로 정리
  */
 @Slf4j
 @Component
 @RequiredArgsConstructor
-public class WeatherChangeDetectionScheduler {
+public class WeatherDataCleanupScheduler {
 
     private final ZoneId zoneId;
 
     @Qualifier("asyncJobLauncher")
     private final JobLauncher asyncJobLauncher;
 
-    @Qualifier("weatherChangeDetectionJob")
-    private final Job weatherChangeDetectionJob;
+    @Qualifier("weatherDataCleanupJob")
+    private final Job weatherDataCleanupJob;
 
-    @Value("${batch.scheduler.weather-change-detection.enabled:true}")
-    private boolean weatherChangeDetectionEnabled;
+    @Value("${batch.scheduler.weather-cleanup.enabled:true}")
+    private boolean weatherCleanupEnabled;
 
     /**
-     * 날씨 변화 감지 작업
-     * 매시 50분에 실행 (초단기예보 제공 후 5분 뒤)
+     * 날씨 데이터 정리 작업
+     * 매일 자정에 실행
      */
-    @Scheduled(cron = "0 50 * * * ?", zone = "#{@timezoneId}")
-    public void runWeatherChangeDetection() {
-        if (!weatherChangeDetectionEnabled) {
-            log.debug("날씨 변화 감지 작업이 비활성화되어 있습니다");
+    @Scheduled(cron = "0 0 0 * * ?", zone = "#{@timezoneId}")
+    public void runWeatherDataCleanup() {
+        if (!weatherCleanupEnabled) {
+            log.debug("날씨 데이터 정리 작업이 비활성화되어 있습니다");
             return;
         }
 
-        log.info("[Scheduled] 날씨 변화 감지 작업 시작");
+        log.info("[Scheduled] 날씨 데이터 정리 작업 시작");
 
         try {
             JobParameters jobParameters = new JobParametersBuilder()
                 .addLong("timestamp", System.currentTimeMillis())
-                .addString("jobType", "scheduled_weather_change_detection")
+                .addString("jobType", "scheduled_weather_cleanup")
                 .addString("triggeredBy", "scheduler")
                 .addString("timezone", zoneId.getId())
                 .toJobParameters();
 
-            JobExecution jobExecution = asyncJobLauncher.run(weatherChangeDetectionJob, jobParameters);
+            JobExecution jobExecution = asyncJobLauncher.run(weatherDataCleanupJob, jobParameters);
 
-            log.info("[Scheduled] 날씨 변화 감지 작업 시작 - Job ID={}, Status={}",
+            log.info("[Scheduled] 날씨 데이터 정리 작업 시작 - Job ID={}, Status={}",
                 jobExecution.getId(), jobExecution.getStatus());
 
         } catch (Exception e) {
-            log.error("[Scheduled] 날씨 변화 감지 작업 실행 실패", e);
+            log.error("[Scheduled] 날씨 데이터 정리 작업 실행 실패", e);
         }
     }
 }
